Move long running test into test without timeout, fixes notnoop/java-apns#145
