+++ src/test/java/com/notnoop/apns/integration/ApnsSimulatorTest.java
@@ -1,79 +1,36 @@
 package com.notnoop.apns.integration;
 
-import com.notnoop.apns.*;
-import com.notnoop.apns.internal.Utilities;
-import com.notnoop.apns.utils.FixedCertificates;
-import com.notnoop.apns.utils.Simulator.FailingApnsServerSimulator;
-import org.hamcrest.BaseMatcher;
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-import org.junit.*;
-import org.junit.rules.TestName;
+import com.notnoop.apns.DeliveryError;
+import org.junit.Assert;
+import org.junit.Ignore;
+import org.junit.Rule;
+import org.junit.Test;
 import org.junit.rules.Timeout;
 import org.mockito.Matchers;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import uk.org.lidalia.slf4jext.Level;
 import uk.org.lidalia.slf4jtest.LoggingEvent;
 import uk.org.lidalia.slf4jtest.TestLoggerFactory;
 
 import java.util.List;
-import java.util.Random;
 import java.util.concurrent.TimeUnit;
 
-import static com.notnoop.apns.utils.FixedCertificates.LOCALHOST;
-import static com.notnoop.apns.utils.FixedCertificates.clientContext;
-import static org.hamcrest.CoreMatchers.*;
+import static org.hamcrest.CoreMatchers.hasItem;
+import static org.hamcrest.CoreMatchers.not;
 import static org.junit.Assert.assertThat;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 @SuppressWarnings("deprecation")
-public class ApnsSimulatorTest {
+public class ApnsSimulatorTest extends ApnsSimulatorTestBase {
 
-    final Logger logger = LoggerFactory.getLogger(ApnsSimulatorTest.class);
+    // final Logger logger = LoggerFactory.getLogger(ApnsSimulatorTest.class);
 
     //@Rule
     //public DumpThreadsOnErrorRule dump = new DumpThreadsOnErrorRule();
 
     @Rule
-    public TestName name = new TestName();
+    public Timeout timeout = new Timeout(5000);
 
-    @Rule
-    public Timeout globalTimeout = new Timeout(5000);
-
-
-    private static final String payload = "{\"aps\":{}}";
-    private ApnsService service;
-    private FailingApnsServerSimulator server;
-    private ApnsDelegate delegate;
-    private Random random;
-
-    @Before
-    public void startup() {
-        // http://projects.lidalia.org.uk/slf4j-test/
-        TestLoggerFactory.getInstance().setPrintLevel(Level.DEBUG);
-        TestLoggerFactory.clearAll();
-
-        logger.info("\n\n\n\n\n");
-        logger.info("********* Test: {}", name.getMethodName());
-
-        server = new FailingApnsServerSimulator(FixedCertificates.serverContext().getServerSocketFactory());
-        server.start();
-        delegate = ApnsDelegate.EMPTY;
-        delegate = mock(ApnsDelegate.class);
-        service = APNS.newService()
-                .withSSLContext(clientContext())
-                .withGatewayDestination(LOCALHOST, server.getEffectiveGatewayPort())
-                .withFeedbackDestination(LOCALHOST, server.getEffectiveFeedbackPort())
-                .withDelegate(delegate).build();
-        random = new Random();
-    }
-
-    @After
-    public void tearDown() {
-        server.stop();
-        server = null;
-    }
 
     @Test
     public void sendOne() throws InterruptedException {
@@ -154,119 +111,10 @@ public class ApnsSimulatorTest {
         assertThat(allLoggingEvents, not(hasItem(eventContains("Exception while waiting for error code"))));
     }
 
-    @Ignore("fails")
     @Test
     public void firstTokenBad_issue145() throws InterruptedException {
         // Test for Issue #145
-        send(8, -1);
+        send(8, 0);
         assertNumberReceived(2);
-    }
-
-    @Ignore("fails")
-    @Test
-    public void multipleTokensBad_issue145() throws InterruptedException {
-        send(8, 0, 8, 0, 8, 0 ,8, 0);
-        assertNumberReceived(8);
-
-    }
-
-    private Matcher<? super LoggingEvent> eventContains(final String substr) {
-        return new BaseMatcher<LoggingEvent>() {
-            @Override
-            public boolean matches(final Object item) {
-                final String message = ((LoggingEvent) item).getMessage();
-                return message.contains(substr);
-            }
-
-            @Override
-            public void describeTo(final Description description) {
-                description.appendText("substring [" + substr + "]");
-            }
-        };
-    }
-
-    private void send(final int... codes) {
-        for (int code : codes) {
-            send(code);
-        }
-    }
-
-    private void send(final int code) {
-
-        ApnsNotification notification = makeNotification(code);
-        service.push(notification);
-
-    }
-
-    /**
-     * Create an APNS notification that creates specified "error" behaviour in the
-     * {@link FailingApnsServerSimulator}
-     *
-     * @param code A code specifying the FailingApnsServer's behaviour.
-     *             <ul>
-     *             <li>-100: Drop connection</li>
-     *             <li>below zero: wait (-code) number times a tenth of a second</li>
-     *             <li>above zero: send code as APNS error message then drop connection</li>
-     *             </ul>
-     * @return an APNS notification
-     */
-    private EnhancedApnsNotification makeNotification(final int code) {
-        byte[] deviceToken = new byte[32];
-        random.nextBytes(deviceToken);
-        if (code == 0) {
-            deviceToken[0] = 42;
-        } else {
-            deviceToken[0] = (byte) 0xff;
-            deviceToken[1] = (byte) 0xff;
-
-            if (code < -100) {
-                // Drop connection
-                deviceToken[2] = (byte) 2;
-            } else if (code < 0) {
-                // Sleep
-                deviceToken[2] = (byte) 1;
-                deviceToken[3] = (byte) -code;
-            } else {
-                // Send APNS error response then drop connection
-                assert code > 0;
-                deviceToken[2] = (byte) 0;
-                deviceToken[3] = (byte) code;
-
-            }
-
-        }
-        return new EnhancedApnsNotification(EnhancedApnsNotification.INCREMENT_ID(), 1, deviceToken, Utilities.toUTF8Bytes(payload));
-    }
-
-    private void sendCount(final int count, final int code) {
-        for (int i = 0; i < count; ++i) {
-            send(code);
-        }
-    }
-
-    private void assertNumberReceived(final int count) throws InterruptedException {
-        logger.debug("assertNumberReceived {}", count);
-        int i = 0;
-        try {
-            for (; i < count; ++i) {
-                server.getQueue().poll(5, TimeUnit.SECONDS);
-            }
-        } catch (RuntimeException re) {
-            logger.error("Exception in assertNumberReceived, took {}", i);
-            throw re;
-        }
-        logger.debug("assertNumberReceived - successfully took {}", count);
-        assertIdle();
-    }
-
-    private void assertIdle() throws InterruptedException {
-        logger.info("assertIdle");
-        Thread.sleep(1000);
-        assertThat(server.getQueue().size(), equalTo(0));
-    }
-
-    private void assertDelegateSentCount(final int count) {
-        logger.info("assertDelegateSentCount {}", count);
-        verify(delegate, times(count)).messageSent(Matchers.any(ApnsNotification.class), Matchers.anyBoolean());
     }
 }
